DeleteTherapistREST enhanced, check if therapist has dependent patients

+++ src/main/java/org/mentalizr/backend/rest/endpoints/admin/userManagement/therapist/DeleteTherapistREST.java
@@ -5,11 +5,14 @@ import de.arthurpicht.webAccessControl.auth.Authorization;
 import de.arthurpicht.webAccessControl.auth.UnauthorizedException;
 import org.mentalizr.backend.accessControl.roles.Admin;
 import org.mentalizr.backend.rest.service.Service;
+import org.mentalizr.persistence.rdbms.barnacle.connectionManager.BusinessConstraintException;
 import org.mentalizr.persistence.rdbms.barnacle.connectionManager.DataSourceException;
 import org.mentalizr.persistence.rdbms.barnacle.connectionManager.EntityNotFoundException;
 import org.mentalizr.persistence.rdbms.barnacle.dao.RoleTherapistDAO;
 import org.mentalizr.persistence.rdbms.barnacle.dao.UserDAO;
 import org.mentalizr.persistence.rdbms.barnacle.dao.UserLoginDAO;
+import org.mentalizr.persistence.rdbms.barnacle.vo.RolePatientVO;
+import org.mentalizr.persistence.rdbms.barnacle.vo.RoleTherapistVO;
 import org.mentalizr.persistence.rdbms.barnacle.vo.UserLoginVO;
 import org.mentalizr.persistence.rdbms.edao.PolicyConsentEDAO;
 
@@ -18,6 +21,7 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import java.util.List;
 
 @Path("v1")
 public class DeleteTherapistREST {
@@ -46,8 +50,13 @@ public class DeleteTherapistREST {
             }
 
             @Override
-            protected Object workLoad() throws DataSourceException, EntityNotFoundException {
+            protected Object workLoad() throws DataSourceException, EntityNotFoundException, BusinessConstraintException {
                 UserLoginVO userLoginVO = UserLoginDAO.findByUk_username(username);
+                RoleTherapistVO roleTherapistVO = RoleTherapistDAO.load(userLoginVO.getUserId());
+
+                List<RolePatientVO> rolePatientVOList = roleTherapistVO.getRolePatientVOByFk_therapist_id();
+                if (!rolePatientVOList.isEmpty())
+                    throw new BusinessConstraintException("Therapist has dependent patients.");
 
                 RoleTherapistDAO.delete(userLoginVO.getUserId());
                 UserLoginDAO.delete(userLoginVO.getUserId());
